python_libraries/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via aiohttp
beautifulsoup4==4.13.4
    # via -r requirements.in
black==22.12.0
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via black
coverage[toml]==7.7.0
    # via pytest-cov
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
docopt==0.6.2
    # via -r requirements.in
duckdb==1.2.1
    # via -r requirements.in
faker==37.0.0
    # via -r requirements.in
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
idna==2.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
jinja2==2.11.3
    # via -r requirements.in
jiter==0.9.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.25
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.23
    # via -r requirements.in
langchain-core==0.3.58
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.16
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via
    #   -r requirements.in
    #   langchain
langsmith==0.3.42
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==1.1.1
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
numpy==2.2.5
    # via langchain-community
openai==1.78.0
    # via langchain-openai
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via -r requirements.in
pydantic==2.11.4
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via langchain-community
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.25.3
    # via -r requirements.in
pytest-cov==6.0.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
redis==5.2.1
    # via -r requirements.in
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.40
    # via
    #   langchain
    #   langchain-community
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.9.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.1
    # via faker
urllib3==2.4.0
    # via requests
yarl==1.20.0
    # via aiohttp
zstandard==0.23.0
    # via langsmith
